+++ code/Samples/MultipleRenderTarget/MultipleRenderTarget.cc
@@ -212,8 +212,8 @@ MultipleRenderTargetApp::notSupported() {
     #else
     const char* msg = "This demo needs MultipleRenderTarget\n";
     #endif
-    uint8_t x = (Gfx::DisplayAttrs().FramebufferWidth/16 - strlen(msg))/2;
-    uint8_t y = Gfx::DisplayAttrs().FramebufferHeight/16/2;
+    uint8_t x = uint8_t((Gfx::DisplayAttrs().FramebufferWidth/16 - strlen(msg))/2);
+    uint8_t y = uint8_t(Gfx::DisplayAttrs().FramebufferHeight/16/2);
     Gfx::BeginPass(PassState(glm::vec4(0.5f, 0.0f, 0.0f, 1.0f)));
     Dbg::SetTextScale(glm::vec2(2.0f, 2.0f));
     Dbg::CursorPos(x, y);
